refactor(hexa): Extract application from core
