+++ application/jest.config.js
@@ -2,7 +2,8 @@ module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
     moduleNameMapper: {
-      '^application/(.*)$': '<rootDir>/application/$1',
+      '^repositories/(.*)$': '<rootDir>/repositories/$1',
+      '^usecases/(.*)$': '<rootDir>/usecases/$1',
       '^domain/(.*)$': '<rootDir>/domain/$1',
     },
     roots: ['<rootDir>/test'], // Asegúrate de que Jest busque en esta carpeta
