+++ application/test/post/delete.test.ts
@@ -1,7 +1,7 @@
 // test/post/DeletePost.test.ts
 
-import { PostRepository } from "application/repositories/post";
-import { DeletePost } from "application/usecases/atomic/post";
+import { PostRepository } from "repositories/post";
+import { DeletePost } from "usecases/atomic/post";
