+++ application/test/user/create.test.ts
@@ -1,5 +1,5 @@
-import { UserRepository } from "application/repositories/user";
-import { CreateUser } from "application/usecases/atomic/user";
+import { UserRepository } from "repositories/user";
+import { CreateUser } from "usecases/atomic/user";
 import { User } from "../../../domain/entities/User";
