+++ application/test/user/read-all.test.ts
@@ -1,6 +1,7 @@
-import { UserRepository } from "application/repositories/user";
-import { ReadAll } from "application/usecases/atomic/user";
+
+import { UserRepository } from "repositories/user";
 import { User } from "../../../domain/entities/User";
+import { ReadAll } from "usecases/atomic/user";
 
 // Mock del repositorio
 const mockUserRepository = (): jest.Mocked<UserRepository> => ({
