+++ infrastructure/repositories/prisma-likepost.ts
@@ -1,6 +1,6 @@
 import { PrismaClientConfig } from "../connectors/prisma-db";
-import { LikePostRepository } from "../../core/application/repositories/likepost";
 import { LikePost } from "../../domain/entities/LikePost";
+import { LikePostRepository } from "../../application/repositories/likepost";
 
 export class PrismaLikePostRepository extends PrismaClientConfig implements LikePostRepository {
     async create(postId: number, userId: number): Promise<LikePost> {
