+++ infrastructure/repositories/prisma-post.ts
@@ -1,6 +1,6 @@
 import { PrismaClientConfig } from "../connectors/prisma-db";
 import { Post } from "../../domain/entities/Post"
-import { PostRepository } from "../../core/application/repositories/post"
+import { PostRepository } from "../../application/repositories/post";
 
 export class PrismaPostRepository extends PrismaClientConfig implements PostRepository {
