+++ infrastructure/repositories/prisma-user.ts
@@ -1,9 +1,9 @@
 // backend/src/infrastructure/repositories/PrismaUserRepository.ts
 
 import { Prisma } from '@prisma/client';
-import {UserRepository}from "../../core/application/repositories/user"
 import { PrismaClientConfig } from '../connectors/prisma-db';
 import { User } from '../../domain/entities/User';
+import { UserRepository } from '../../application/repositories/user';
 
 export class PrismaUserRepository extends PrismaClientConfig implements UserRepository {
   constructor() {
