+++ infrastructure/tsconfig.json
@@ -9,6 +9,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["./prisma", "./repositories", "./connectors", "../domain"],
+  "include": ["./prisma", "./repositories", "./connectors", "../domain", "../application"],
   "exclude": ["node_modules"]
 }
